[PI-230] Refactoring and renaming
